% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_df_2_sample_t_test.R
\name{approx_df_2_sample_t_test}
\alias{approx_df_2_sample_t_test}
\title{Gets the degrees of freedom in 2-samp t-test}
\usage{
approx_df_2_sample_t_test(x, y)
}
\arguments{
\item{x}{data for sample 1}

\item{y}{data for sample 2}
}
\value{
degrees of freedom
}
\description{
This function takes in 2 vectors of a 2-sample t-test and
returns the degrees of freedom
}
\author{
Timothy Roeper, PhD Professor at New York University
}
