% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypotest.R
\name{get_p_value}
\alias{get_p_value}
\title{Gets p-value}
\usage{
get_p_value(z_score, two_sided = FALSE)
}
\arguments{
\item{z_score}{Z-score}

\item{two_sided}{Boolean value denoting whether test is double sided or not, defaults to FALSE}
}
\value{
p-value
}
\description{
Defaults to 1 sided test, call with parameter (doubleSided = TRUE) to get p-value of a double sided test
}
