% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypotest.R
\name{get_z}
\alias{get_z}
\title{Get Z score}
\usage{
get_z(p_hat, p_null, size, sigma = NULL)
}
\arguments{
\item{p_hat}{Sample probability}

\item{p_null}{null hypothesis value}

\item{size}{Sample size}

\item{sigma}{(Optional) Standard deviation, defaults to NULL}
}
\value{
Z-score
}
\description{
This function gets z-score given sample p (p_hat), null hypothesis value (p_null), and sample size
}
