% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypotest.R
\name{get_z_given_sd}
\alias{get_z_given_sd}
\title{Gets Z-score given standard deviation. You can just call get_z() and pass in the standard deviation.}
\usage{
get_z_given_sd(p_hat, p_null, sd, size)
}
\arguments{
\item{p_hat}{Sample probability}

\item{sd}{Standard deviation}

\item{size}{Sample size}

\item{p_sull}{Null hypothesis value}
}
\value{
Z-score
}
\description{
Gets z-score given sample p (p_hat), null hypothesis value (p_null), population standard deviation (sd), and sample
}
