% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypotest.R
\name{getP}
\alias{getP}
\title{Gets p-value}
\usage{
getP(zScore, doubleSided = FALSE)
}
\arguments{
\item{zScore}{z-score}

\item{doubleSided}{Boolean value denoting whether test is double sided or not, defaults to FALSE}
}
\value{
p-value
}
\description{
usage: defaults to 1 sided test, call with parameter (doubleSided = TRUE) to get p-value of a double sided test
}
