% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypotest.R
\name{getZ}
\alias{getZ}
\title{Get Z score}
\usage{
getZ(pHat, pNull, size)
}
\arguments{
\item{pHat}{Sample probability}

\item{pNull}{null hypothesis value}

\item{size}{Sample size}
}
\value{
Z-score
}
\description{
This function gets z-score given sample p (pHat), null hypothesis value (pNull), and sample size
}
