% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypotest.R
\name{getZGivenSD}
\alias{getZGivenSD}
\title{Gets Z-score given standard deviation}
\usage{
getZGivenSD(pHat, pNull, sd, size)
}
\arguments{
\item{pHat}{Sample probability}

\item{pNull}{Null hypothesis value}

\item{sd}{Standard deviation}

\item{size}{Sample size}
}
\value{
Z-score
}
\description{
gets z-score given sample p (pHat), null hypothesis value (pNull), population standard deviation (sd), and sample
}
