% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence-interval.R
\name{interval}
\alias{interval}
\title{Get confidence interval}
\usage{
interval(trials, successes, conf)
}
\arguments{
\item{trials}{Number of trials}

\item{successes}{Number of successes of the trials}

\item{conf}{Confidence level}
}
\value{
Confidence level formatted as a String
}
\description{
This function gets the confidence interval given the
number of trials, number of successes, and confidence level.
Confidence level should be given in decimal form, not percentage.
}
