% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{linear_regression}
\alias{linear_regression}
\title{Returns linear regression model (least squares), sum of residuals squared, total resioduals squared,
y-hat predicted values, residuals squared, total squares given vectors x, y}
\usage{
linear_regression(x, y)
}
\arguments{
\item{vector}{y for dependent varaible values}
}
\value{
data frame containing linear regression results, including R^2.
}
\description{
Returns linear regression model (least squares), sum of residuals squared, total resioduals squared,
y-hat predicted values, residuals squared, total squares given vectors x, y
}
