% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power}
\alias{power}
\title{Gets powe, or the probability of correctly rejecting the null hypothesis.}
\usage{
power(Ho, Ha, mu_alt, n, alpha, sigma = NULL)
}
\arguments{
\item{Ho}{Null hypothesis}

\item{Ha}{Direction of the test, valid arguments are either "<", ">", or "!="}

\item{mu_alt}{True mean or probability}

\item{n}{Sample size}

\item{alpha}{Significance level/Type I error probability/alpha level}

\item{sigma}{(Optional) population standard deviation}
}
\value{
probability of Type II error
}
\description{
This function will return the probability of rejecting the null hypothesis
when it is in fact false, given the null hypothesis value, direction of the
test, alternative mu value, sample size, significane level alpha, and
(optionally) the population standard deviation
}
