% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_tests.R
\name{two_samp_binom}
\alias{two_samp_binom}
\title{2-sample z-test with binomial distribution}
\usage{
two_samp_binom(
  p1_trials,
  p1_successes,
  p2_trials,
  p2_successes,
  H0_diff = 0,
  Ha,
  alpha = 0.05
)
}
\arguments{
\item{p1_trials}{number of trials in the first group}

\item{p1_successes}{number of successes in the first group}

\item{p2_trials}{number of trails in the second group}

\item{p2_successes}{number of successes in the second group}

\item{H0_diff}{H0 value that tests the difference, defaults to 0 (testing if two samples have equal means}

\item{Ha}{direction of the test. Accepts either ">", "<", or "!="}

\item{alpha}{significance level, or the probability of making a type 1 error. Defaults to 0.05}
}
\value{
data frame containing p-value, z-score, and result of the hypothesis test
}
\description{
Given that the samples use a binomial distribution, this function performs a 2-sample z-test
}
