% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_tests.R
\name{two_samp_t_test_neq}
\alias{two_samp_t_test_neq}
\title{Performs 2-sided 2-sample t-test}
\usage{
two_samp_t_test_neq(
  samp1_mean,
  samp2_mean,
  samp1_size,
  samp2_size,
  s1,
  s2,
  H0_diff = 0,
  alpha = 0.05,
  df = NULL
)
}
\arguments{
\item{samp1_mean}{mean of the first group}

\item{samp2_mean}{mean of the first group}

\item{samp1_size}{mean of the first group}

\item{samp2_size}{mean of the first group}

\item{s1}{sample standard deviation of the first group}

\item{s2}{sample standard deviation of the second group}

\item{H0_diff}{H0 value that tests the difference, defaults to 0 (testing if two samples have equal means}

\item{alpha}{significance level, or the probability of making a type 1 error. Defaults to 0.05}

\item{df}{degrees of freedom, defaults to NULL (program will calculate it if not specified)}

\item{Ha}{direction of the test. Accepts either ">", "<", or "!="}
}
\value{
data frame containing p-value, z-score, and result of the hypothesis test
}
\description{
Tests the alternate hypothesis that states the H0 value is not H0_diff (usually 0),
resulting in a 2-sided test
}
