% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_tests.R
\name{two_samp_test_equal_var}
\alias{two_samp_test_equal_var}
\title{Runs a 2-sample t test assuming equal variances}
\usage{
two_samp_test_equal_var(samp1, samp2, H0_diff = 0, Ha, alpha = 0.05)
}
\arguments{
\item{samp1}{data for the first sample}

\item{samp2}{data for the second sample}

\item{H0_diff}{H0 value that tests the difference, defaults to 0 (testing if two samples have equal means}

\item{Ha}{direction of the test. Accepts either ">", "<", or "!="}

\item{alpha}{significance level, or the probability of making a type 1 error. Defaults to 0.05}
}
\value{
data frame containing p-value and result of the hypothesis test
}
\description{
Given that the variances are equal, the degrees of freedom calculation is much simpler. This
hypotest formula takes advantage of the it and performs a 2-sample t-test
}
