% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type2.R
\name{type2err}
\alias{type2err}
\title{Generates probability of Type II error}
\usage{
type2err(Ho, Ha, p, n, alpha)
}
\arguments{
\item{Ho}{Null hypothesis}

\item{Ha}{Direction of the test, valid arguments are either "<", ">", or "!="}

\item{p}{True probability}

\item{n}{Sample size}

\item{alpha}{Significance level/Type I error probability/alpha level}
}
\value{
probability of Type II error
}
\description{
This function generates the probability of having a type II error given
the null hypothesis, what the alternate hypothesis tests (higher, lower, or not equals to),
the actual probability/mean, sample size, and significance level alpha
}
