% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypotest.R
\name{z_test_leq}
\alias{z_test_leq}
\title{Conducts lower-tailed z-test}
\usage{
z_test_leq(H0, sigma = NULL, alpha = 0.05, samp_size, x_bar)
}
\arguments{
\item{H0}{Null hypothesis}

\item{sigma}{(Optional) Population standard deviation, defaults to NULL}

\item{alpha}{Significance level, defaults to 0.05 or 5\%}

\item{samp_size}{Sample size}

\item{x_bar}{Sample mean or proportion}
}
\value{
data frame containing test result, p-value, and critical value
}
\description{
Gets the outcome, p-value, and upper critical value of a lower-tailed hypothesis test (z-test)
given the null hypothesis, population standard deviation, significance leve, sample size, and sample mean
}
